fixes #2560 Cpu speed is now an Integer in ldap, search on that field are now working

+++ inventory-api/src/main/scala/com/normation/inventory/domain/MachineInventory.scala
@@ -81,7 +81,7 @@ case class Port(
 case class Processor(
   name:String,
   description:Option[String] = None,
-  speed : Option[Float] = None,
+  speed : Option[Int] = None,
   stepping : Option[Int] = None,
   model : Option[String] = None, 
   family : Option[String] = None,
